将MacroBiome重写 添加河谷侵蚀系统 准备参数化河谷侵蚀

+++ src/main/java/com/EyeOfHarmonyBuffer/space/talos/chunk/coastline/CoastlineSample.java
@@ -3,11 +3,11 @@ package com.EyeOfHarmonyBuffer.space.talos.chunk.coastline;
 public final class CoastlineSample {
 
     private final boolean land;
-    private final int distanceToCoast;
+    private final double distanceToCoast;
     private final int beachWidth;
     private final int shelfWidth;
 
-    public CoastlineSample(boolean land, int distanceToCoast, int beachWidth, int shelfWidth) {
+    public CoastlineSample(boolean land, double distanceToCoast, int beachWidth, int shelfWidth) {
         this.land = land;
         this.distanceToCoast = distanceToCoast;
         this.beachWidth = beachWidth;
@@ -15,7 +15,7 @@ public final class CoastlineSample {
     }
 
     public boolean isLand() { return land; }
-    public int distanceToCoast() { return distanceToCoast; }
+    public double distanceToCoast() { return distanceToCoast; }
     public int beachWidth() { return beachWidth; }
     public int shelfWidth() { return shelfWidth; }
 }
